Rename CardReaderScanFragment to CardReaderConnectFragment
