+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/connect/CardReaderConnectFragment.kt
@@ -1,0 +1,16 @@
+package com.woocommerce.android.ui.prefs.cardreader.connect
+
+import android.os.Bundle
+import android.view.View
+import com.woocommerce.android.R
+import com.woocommerce.android.databinding.CardReaderConnectBinding
+import com.woocommerce.android.ui.base.BaseFragment
+
+class CardReaderConnectFragment : BaseFragment(R.layout.card_reader_connect) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val binding = CardReaderConnectBinding.bind(view)
+        binding.testingText.setText("Hardcoded: Card Reader Connect Fragment")
+    }
+}
