+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/connect/CardReaderConnectViewModel.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.prefs.cardreader.scan
+package com.woocommerce.android.ui.prefs.cardreader.connect
 
 import com.woocommerce.android.di.ViewModelAssistedFactory
 import com.woocommerce.android.util.CoroutineDispatchers
@@ -10,15 +10,15 @@ import dagger.assisted.AssistedInject
 import org.wordpress.android.fluxc.utils.AppLogWrapper
 import org.wordpress.android.util.AppLog.T
 
-class CardReaderScanViewModel @AssistedInject constructor(
+class CardReaderConnectViewModel @AssistedInject constructor(
     @Assisted savedState: SavedStateWithArgs,
     dispatchers: CoroutineDispatchers,
     private val appLogWrapper: AppLogWrapper
 ) : DaggerScopedViewModel(savedState, dispatchers) {
     fun foo() {
-        appLogWrapper.d(T.MAIN, "Scan foo() invoked")
+        appLogWrapper.d(T.MAIN, "Connect foo() invoked")
     }
 
     @AssistedFactory
-    interface Factory : ViewModelAssistedFactory<CardReaderScanViewModel>
+    interface Factory : ViewModelAssistedFactory<CardReaderConnectViewModel>
 }
