+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/detail/CardReaderDetailViewModel.kt
@@ -1,7 +1,7 @@
 package com.woocommerce.android.ui.prefs.cardreader.detail
 
 import androidx.lifecycle.SavedStateHandle
-import com.woocommerce.android.ui.prefs.cardreader.detail.CardReaderDetailViewModel.NavigationTarget.CardReaderScanScreen
+import com.woocommerce.android.ui.prefs.cardreader.detail.CardReaderDetailViewModel.NavigationTarget.CardReaderConnectScreen
 import com.woocommerce.android.util.CoroutineDispatchers
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event
 import com.woocommerce.android.viewmodel.ScopedViewModel
@@ -13,11 +13,11 @@ class CardReaderDetailViewModel @Inject constructor(
     savedState: SavedStateHandle,
     dispatchers: CoroutineDispatchers
 ) : ScopedViewModel(savedState, dispatchers) {
-    fun onInitiateScanBtnClicked() {
-        triggerEvent(CardReaderScanScreen)
+    fun onConnectBtnClicked() {
+        triggerEvent(CardReaderConnectScreen)
     }
 
     sealed class NavigationTarget : Event() {
-        object CardReaderScanScreen : NavigationTarget()
+        object CardReaderConnectScreen : NavigationTarget()
     }
 }
