+++ WooCommerce/src/main/res/layout/card_reader_connect.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent">
 
     <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/testing_text_scan"
+        android:id="@+id/testing_text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         app:layout_constraintBottom_toBottomOf="parent"
