+++ WooCommerce/src/main/res/navigation/nav_graph_settings.xml
@@ -53,16 +53,16 @@
             app:popExitAnim="@anim/activity_slide_out_to_right" />
     </fragment>
     <fragment
-        android:id="@+id/cardReaderScanFragment"
-        android:name="com.woocommerce.android.ui.prefs.cardreader.scan.CardReaderScanFragment"
-        android:label="CardReaderScanFragment" />
+        android:id="@+id/cardReaderConnectFragment"
+        android:name="com.woocommerce.android.ui.prefs.cardreader.connect.CardReaderConnectFragment"
+        android:label="CardReaderConnectFragment" />
     <fragment
         android:id="@+id/cardReaderDetailFragment"
         android:name="com.woocommerce.android.ui.prefs.cardreader.detail.CardReaderDetailFragment"
         android:label="CardReaderDetailFragment" >
         <action
-            android:id="@+id/action_cardReaderDetailFragment_to_cardReaderScanFragment"
-            app:destination="@id/cardReaderScanFragment"
+            android:id="@+id/action_cardReaderDetailFragment_to_cardReaderConnectFragment"
+            app:destination="@id/cardReaderConnectFragment"
             app:enterAnim="@anim/activity_slide_in_from_right"
             app:exitAnim="@anim/activity_slide_out_to_left"
             app:popEnterAnim="@anim/activity_slide_in_from_left"
